test(particle): fold orbital velocity coverage into existing suite
